Updated README with the start of Dev IDE Setup
I've started looking into IDEs. The one at this point I'm going to propose to the team that we use is the PlatformIO IDE. I started working on the documentation on getting the IDE installed and set up properly for our project.

I still need to run it through the ringer a little bit to make sure we want to actually use that IDE but on paper, based on the reasons I now have listed in the repo, it seems the most promising. _(CLion is probably the one I would use since I'm so used to IntelliJ but it seems like a bunch more configuration that we'd have to do to get it set up. It isn't as turn-key as PlatformIO.)_