Added Button Class
I realized that we also have a Button component in our base module so I wanted to get a class made for that. I also added a check in main.cpp which checks if it is pressed and sounds a tone on the buzzer if it is.

Also, I added a few methods to the Buzzer class so that you can manually control it instead of just specifying an amount of time to sound it for. If you use the method without the millisOn then it will just indefinately sound until you call the stopTone() method.