Corrected the Photoresistor's Name
I originally named the file "PhotoResistor.cpp" but then I've referenced it as "Photoresistor.cpp" (note the "R" vs. "r" in resistor).

On some platforms / using some compilers it handled this difference just fine. On some it blows chunks...

This change renames the files so that this won't be any issue.