Added a testComponent() method
I've added a method which collapses both the "set" and the "test" methods into a single call. This will help keep our testBoardComponents() method a little shorter.

Note - I left the original pattern for the constructor and the testPin() because we haven't defined a no-argument constructor yet. This would be another good refactor to do.

- Also fixed a small typo