Added a basic Properties class
It is seeming like we are going to need a place to start storing Proptery-like key-value pairs that we can access statically throughout the project. It feels clunky to continue to bloat main.cpp.

While not perfect, this at least gives us a starting point to keep all of our Properties in a single place.