Last bit of initial work on the Wifi Module
This is the last commit of the initial work for the Wifi Module. There is still LOTS to do on it, but this gets us in a good place for this to be merged into master.

The Properties file should be ready to be utilized throughout the project at this point.