Add Logging Documentation
We decided that we need to change the way we are logging these. We discussed as a group a Logging & Command Message scheme we could start to use. This commit adds the documenation around that scheme we came up with.

At this point, none of this documentation has been implemented. We wanted the documenation in place before we started the refactor.

Note - the values currently in the `Details` column of the table are just placeholders and will be updated with actual values as we perform the refactor.