Add Defn. of Execute Commands and added Comment Convention
I've added the definition for the 4 commands currently defined in the Ardunino's code to handle.

I also added a Comment Convention that we should follow as we start working on the refactor for the logging, etc. Following this will help save a little time when a developer is lookng at the code to not have to go look things up in the table and instead can just see it documented right in the code.