Documentation of base classes
Add documentation to all of the functions in the main source library for easy read and to most of the paramters, except for those in the properites class. This is meant to make the code more readible to other devs who pick it up as it goes. Suggesting to keep documentation as the project goes.