Changed from "Chris Nash" to "Christopher Nash"
In print, etc. I generally use my full first name. This change just modifies the comments on many of the files to be `Christopher Nash` instead of `Chris Nash`. This also makes it consistent with the way it is written in the main README file.

There are NO actual code changes in this commit. Only comment changes.