Wifi Module Refactor
This change contains several items related to the Wifi Module.

The first is the addition of the `lib_deps` section in the `platformio.ini` file. This uses PlatformIO's [Library management system](http://platformio.org/lib) to pull in the required `WiFiEsp` library.

The next main refactor was renaming from `Wifi` to `WifiModule`. I had to do this as the WifiEsp library had a constant named Wifi which caused some compile and runtime issues. The quick-win was just renaming our file instead of working around it.

The final refactor was in the wifi code itself. I cleaned out ALOT of the initial code I had done and replaced it with code utilizing the new WiFiEsp library. For right now that code is commented out and I will fully remove it in a future commit.

NOTE - This currently has an issue where everything compiles and runs but the Arduino isn't actually able to communicate with the ESP8266... I suspect it is happening somewhere between the constructors of the `esp8266` SoftwareSerial object and passing that into the ` WiFi.init(&esp8266);` method. It works if I define the SoftwareSerial and do the WiFi.init() in the main class but thats 'icky' and I want to have all of that contained in this class (if possible). I will trouubleshoot this more in the future and either go the 'icky' route to get it functioning or identify where the real issue is and correct it.