Fixed bug causing Wifi Module not to function
After hours of poking around it all came down to one little thing - I had forgot to start the communication of the esp8266 SoftwareSerial... /facepalm

This is now functioning! The next commit will include some cleanup and hooking it up to actually trigger the notification when the alarm is triggered.