Hooked wifi module up to send alert when triggered
The wifi module will now send an alert when the alarm is triggered!

Note - there is some improvement still to be done on this code. I've added a few TODOs. These will be addressed at some point in the future. There is also some additional documentation, both in the Readme and in the code itself, which should be added. This will also happen in the future.