Made wifi setup sound a 'waiting' tone on the alarm
This change make a 'waiting' sound trigger before the wifi module connects to wirless and a 'success' sound trigger when it successfully connects.

While debugging / working on my wifi module I realized that after the alarm goes through its normal setup procesures it turns control over to the wifi module to do its thing. It was difficult to know when / if it had finally connected to wifi (_which seems to vary anywhere from 5 seconds to almost 30 seconds..._) so I added a way that you can hear and see that it is starting the wifi setup as well as hear and see when it successfully connected.

Unfortunately, once I hand over control to the ESP8266's library I am using things are out of my control. I can't, for example, continue sound the tone every second or continue flashing an led. If we were in a place where we could do multi-threaded operations then I could do that. In theory, if I were to spend a whole bunch of time working on my own library, I could send the command to the ESP8266 chip to start connecting, return control back to the Arduino (since the ESP8266 is a microcontroller itself and can handle its own processing) and then come back periodically to check "Hey - are you connected yet?". But thats not the way these libraries for it are configured to work... So I'm out of luck on that.

So while not great, this at least gives _something_ to let the user know that something is happening....