Added Turret Module
This is the start of the Turret Module. This code is still current UNTESTED as I can't get my old Arduino to work with the I2C protocol correctly.

There is more work needed for this module in both the code and to documentation that will need to come in future commits.

The code does build and shouldn't cause any disruptions.